requirements.txt
requests==2.25.1
python-dotenv==1.0.0

pytest==7.4.3
